fix re renders of wallet view on every navigation action
